Add an documentation to the SimpleMultiResource
This commit adds a documentation to SimpleMultiResource as well as
modifies its deploy script to simply deploy it. User journey script has
been added to replace the showcase of SimpleMultiResource usage.

Custom script for running the SimpleMultiResource user journey has been
added.

MultiResource example smart contracts have been moved into their own
folder, so that the documentation of each architecture can be separate
from the main README.
