+++ contracts/MultiResource/SimpleMultiResource.sol
